pyproject.toml
[project]
name = "nonebot-plugin-maimaidx"
version = "1.0.0"
authors = [
  { name="Yuri-YuzuChaN", email="806235364@qq.com" },
]
description = "maimaidx plugin for nonebot2"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "nonebot2<3.0.0,>=2.0.0",
    "nonebot-adapter-onebot<3.0.0,>=2.2.2",
    "nonebot-plugin-apscheduler<1.0.0,>=0.3.0",
    "aiofiles<24.0.0,>=23.2.1",
    "httpx<1.0.0,>=0.23.1",
    "pillow<11.0.0,>=10.0.0",
    "pydantic<2.0.0,>=1.10.13",
    "pyecharts<3.0.0,>=2.0.0",
    "snapshot_phantomjs<1.0.0,>=0.0.1"
]

[project.urls]
"Homepage" = "https://github.com/Yuri-YuzuChaN/nonebot-plugin-maimaidx"
"Bug Tracker" = "https://github.com/Yuri-YuzuChaN/nonebot-plugin-maimaidx/issues"

[tool.setuptools.package-data]
"*" = ["maimaidxhelp.png"]